+++ core/config_manager/ConfigManagerBase.cpp
@@ -2398,11 +2398,9 @@ bool ConfigManagerBase::GetYamlConfigDirUpdate() {
     std::vector<std::string> filepathes;
     std::unordered_map<std::string, int64_t> yamlConfigMTimeMap;
     static std::string localConfigDirPath = AppConfig::GetInstance()->GetLocalUserYamlConfigDirPath();
-    updateFlag |= CheckYamlDirConfigUpdate(localConfigDirPath, false, filepathes, yamlConfigMTimeMap);
-    // TODO: Change serverConfigDirPath to be parallel to localConfigDirPath. Futhermore, create serverConfigDirPath
-    // only when config server is connected.
-    // static std::string serverConfigDirPath = localConfigDirPath + "remote_config" + PATH_SEPARATOR;
-    // updateFlag |= CheckYamlDirConfigUpdate(serverConfigDirPath, true, filepathes, yamlConfigMTimeMap);
+    updateFlag |= CheckYamlDirConfigUpdate(localConfigDirPath, false, filepathes, yamlConfigMTimeMap, true);
+    static std::string serverConfigDirPath = AppConfig::GetInstance()->GetRemoteUserYamlConfigDirPath();
+    updateFlag |= CheckYamlDirConfigUpdate(serverConfigDirPath, true, filepathes, yamlConfigMTimeMap, false);
 
     if (mYamlConfigMTimeMap.size() != yamlConfigMTimeMap.size()) {
         updateFlag = true;
@@ -2439,7 +2437,8 @@ bool ConfigManagerBase::GetYamlConfigDirUpdate() {
 bool ConfigManagerBase::CheckYamlDirConfigUpdate(const std::string& configDirPath,
                                                  bool isRemote,
                                                  std::vector<std::string>& filepathes,
-                                                 std::unordered_map<std::string, int64_t>& yamlConfigMTimeMap) {
+                                                 std::unordered_map<std::string, int64_t>& yamlConfigMTimeMap,
+                                                 bool createIfNotExist) {
     bool updateFlag = false;
     fsutil::Dir configDir(configDirPath);
     if (!configDir.Open()) {
@@ -2447,7 +2446,7 @@ bool ConfigManagerBase::CheckYamlDirConfigUpdate(const std::string& configDirPat
         if (fsutil::Dir::IsEACCES(savedErrno) || fsutil::Dir::IsENOTDIR(savedErrno)
             || fsutil::Dir::IsENOENT(savedErrno)) {
             LOG_DEBUG(sLogger, ("invalid yaml conf dir", configDirPath)("error", ErrnoToString(savedErrno)));
-            if (!Mkdir(configDirPath.c_str())) {
+            if (createIfNotExist && !Mkdir(configDirPath.c_str())) {
                 savedErrno = errno;
                 if (!IsEEXIST(savedErrno)) {
                     LOG_ERROR(sLogger, ("create conf yaml dir failed", configDirPath)("error", strerror(savedErrno)));